requirements.txt
# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0

# Web scraping
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2

# AI and ML
openai==1.90.0

# Video processing
moviepy==1.0.3
pillow==10.1.0

# Audio processing (keep gTTS as fallback)
gtts==2.4.0
pydub==0.25.1

# Data handling
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
uuid==1.30

aiofiles==23.2.1
jinja2==3.1.2
opencv-python==4.8.1.78
numpy==1.24.3
selenium==4.15.2 